Duplicate the BeforeBind interface, deprecate the old ones, use both when persisting new entities.
