+++ src/main/java/org/springframework/data/neo4j/config/ReactiveNeo4jAuditingRegistrar.java
@@ -25,7 +25,7 @@ import org.springframework.data.auditing.config.AuditingBeanDefinitionRegistrarS
 import org.springframework.data.auditing.config.AuditingConfiguration;
 import org.springframework.data.config.ParsingUtils;
 import org.springframework.data.mapping.context.PersistentEntities;
-import org.springframework.data.neo4j.repository.event.ReactiveAuditingBeforeBindCallback;
+import org.springframework.data.neo4j.core.mapping.callback.ReactiveAuditingBeforeBindCallback;
 import org.springframework.util.Assert;
 
 /**
