+++ src/main/java/org/springframework/data/neo4j/core/mapping/callback/ReactiveBeforeBindCallback.java
@@ -13,25 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.data.neo4j.repository.event;
+package org.springframework.data.neo4j.core.mapping.callback;
 
 import org.apiguardian.api.API;
+import org.reactivestreams.Publisher;
 import org.springframework.data.mapping.callback.EntityCallback;
+import org.springframework.data.mapping.callback.ReactiveEntityCallbacks;
 
 /**
  * Entity callback triggered before an Entity is bound to a record (represented by a {@link java.util.Map
  * java.util.Map&lt;String, Object&gt;}).
  *
- * @author Michael J. Simons
  * @param <T> The type of the entity.
+ * @author Michael J. Simons
+ * @soundtrack Iron Maiden - Killers
+ * @see ReactiveEntityCallbacks
  * @since 6.0
- * @soundtrack Bon Jovi - Slippery When Wet
- * @deprecated since 6.0.2, please use {@link org.springframework.data.neo4j.core.mapping.callback.BeforeBindCallback}.
  */
 @FunctionalInterface
-@API(status = API.Status.DEPRECATED, since = "6.0")
-@Deprecated
-public interface BeforeBindCallback<T> extends EntityCallback<T> {
+@API(status = API.Status.STABLE, since = "6.0")
+public interface ReactiveBeforeBindCallback<T> extends EntityCallback<T> {
 
 	/**
 	 * Entity callback method invoked before a domain object is saved. Can return either the same or a modified instance
@@ -41,5 +42,5 @@ public interface BeforeBindCallback<T> extends EntityCallback<T> {
 	 * @param entity the domain object to save.
 	 * @return the domain object to be persisted.
 	 */
-	T onBeforeBind(T entity);
+	Publisher<T> onBeforeBind(T entity);
 }
